+++ backend/api/ml_model/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ['endpoints', 'predictor']
